+++ examples/resources/hcloud_floating_ip_assignment/resource.tf
@@ -7,7 +7,7 @@ resource "hcloud_server" "node1" {
   name        = "node1"
   image       = "debian-12"
   server_type = "cx23"
-  datacenter  = "fsn1-dc8"
+  location    = "fsn1"
 }
 
 resource "hcloud_floating_ip" "master" {