+++ examples/resources/hcloud_primary_ip/import-by-string-id.tf
@@ -1,0 +1,4 @@
+import {
+  to = hcloud_primary_ip.example
+  id = "$PRIMARY_IP_ID"
+}