+++ examples/resources/hcloud_primary_ip/resource.tf
@@ -1,6 +1,6 @@
 resource "hcloud_primary_ip" "main" {
   name          = "primary_ip_test"
-  datacenter    = "fsn1-dc14"
+  location      = "fsn1"
   type          = "ipv4"
   assignee_type = "server"
   auto_delete   = true
@@ -13,7 +13,7 @@ resource "hcloud_server" "server_test" {
   name        = "test-server"
   image       = "ubuntu-24.04"
   server_type = "cx23"
-  datacenter  = "fsn1-dc14"
+  location    = "fsn1"
   labels = {
     "test" : "tessst1"
   }