+++ examples/resources/hcloud_volume_attachment/resource.tf
@@ -8,7 +8,7 @@ resource "hcloud_server" "node1" {
   name        = "node1"
   image       = "debian-12"
   server_type = "cx23"
-  datacenter  = "nbg1-dc3"
+  location    = "nbg1"
 }
 
 resource "hcloud_volume" "master" {