+++ internal/zone/data_source_test.go
@@ -66,7 +66,7 @@ func TestAccZoneDataSource(t *testing.T) {
 					resource.TestCheckResourceAttr(byName.TFID(), "primary_nameservers.#", "0"),
 					resource.TestCheckResourceAttr(byName.TFID(), "delete_protection", "false"),
 					resource.TestCheckResourceAttr(byName.TFID(), "authoritative_nameservers.assigned.#", "3"),
-					resource.TestCheckResourceAttr(byName.TFID(), "registrar", "other"),
+					resource.TestMatchResourceAttr(byName.TFID(), "registrar", zone.RegistrarRegexp),
 
 					resource.TestCheckResourceAttr(byID.TFID(), "name", res.Name),
 					resource.TestCheckResourceAttr(byID.TFID(), "mode", "primary"),
@@ -75,7 +75,7 @@ func TestAccZoneDataSource(t *testing.T) {
 					resource.TestCheckResourceAttr(byID.TFID(), "primary_nameservers.#", "0"),
 					resource.TestCheckResourceAttr(byID.TFID(), "delete_protection", "false"),
 					resource.TestCheckResourceAttr(byID.TFID(), "authoritative_nameservers.assigned.#", "3"),
-					resource.TestCheckResourceAttr(byID.TFID(), "registrar", "other"),
+					resource.TestMatchResourceAttr(byID.TFID(), "registrar", zone.RegistrarRegexp),
 
 					resource.TestCheckResourceAttr(byLabel.TFID(), "name", res.Name),
 					resource.TestCheckResourceAttr(byLabel.TFID(), "mode", "primary"),
@@ -84,7 +84,7 @@ func TestAccZoneDataSource(t *testing.T) {
 					resource.TestCheckResourceAttr(byLabel.TFID(), "primary_nameservers.#", "0"),
 					resource.TestCheckResourceAttr(byLabel.TFID(), "delete_protection", "false"),
 					resource.TestCheckResourceAttr(byLabel.TFID(), "authoritative_nameservers.assigned.#", "3"),
-					resource.TestCheckResourceAttr(byLabel.TFID(), "registrar", "other"),
+					resource.TestMatchResourceAttr(byLabel.TFID(), "registrar", zone.RegistrarRegexp),
 				),
 			},
 		},